feat: remove lodash dependency, remove @types/mathjs target node 10 BREAKING CHANGE: requires node > 10 BREAKING CHANGE: Trueskill.rate function now only allows one "user" per object if using named ratings, probably no impact
