+++ package.json
@@ -13,7 +13,7 @@
   "sideEffects": false,
   "license": "MIT",
   "scripts": {
-    "benchmark": "ts-node -P ./benchmark/tsconfig.json benchmark/index.ts",
+    "benchmark": "npx ts-node -P ./benchmark/tsconfig.json benchmark/index.ts",
     "lint": "eslint --ext .js,.ts, .",
     "lint:fix": "eslint --fix --ext .js,.ts, .",
     "prepare": "npm run build",
@@ -25,22 +25,21 @@
     "semantic-release": "semantic-release"
   },
   "dependencies": {
-    "@types/mathjs": "6.0.5",
     "mathjs": "^6.6.4",
-    "ts-gaussian": "^1.2.4",
+    "ts-gaussian": "^2.0.0",
     "uuid": "^7.0.3"
   },
   "devDependencies": {
     "@types/benchmark": "1.0.31",
-    "@types/fs-extra": "8.1.0",
     "@types/jest": "25.2.1",
     "@types/lodash": "4.14.150",
+    "@types/mathjs": "6.0.5",
     "@types/node": "13.13.2",
     "@types/uuid": "7.0.2",
     "@typescript-eslint/eslint-plugin": "2.29.0",
     "@typescript-eslint/parser": "2.29.0",
     "benchmark": "2.1.4",
-    "del-cli": "^3.0.0",
+    "del-cli": "3.0.0",
     "eslint": "6.8.0",
     "eslint-config-prettier": "6.11.0",
     "eslint-config-xo-space": "0.24.0",
@@ -50,7 +49,6 @@
     "lodash": "4.17.15",
     "semantic-release": "17.0.7",
     "ts-jest": "25.4.0",
-    "ts-node": "8.9.0",
     "typedoc": "0.17.4",
     "typescript": "3.8.3"
   },
